Remove CapacityByDates table and AdressByDates table.

+++ BatteryCondition/Models/CapacityByDate.cs
@@ -8,8 +8,7 @@ namespace BatteryCondition.Models
     public class CapacityByDate
     {
         public int CapacityByDateId { get; set; }
-        public int Capacity { get; set; }
+        public float Capacity { get; set; }
         public DateTime DateTime { get; set; }
-
     }
 }
